add solutions hook draft
