+++ app/layout/scenarios/sidebar/solutions/details/component.tsx
@@ -5,18 +5,13 @@ import { motion } from 'framer-motion';
 import Icon from 'components/icon';
 import Button from 'components/button';
 import Modal from 'components/modal';
-import Checkbox from 'components/forms/checkbox';
 
 import ARROW_LEFT_SVG from 'svgs/ui/arrow-right-2.svg?sprite';
 import CLOCK_SVG from 'svgs/ui/clock.svg?sprite';
 import TABLE_SVG from 'svgs/ui/table.svg?sprite';
-import INFO_SVG from 'svgs/ui/info.svg?sprite';
-import DOWNLOAD_SVG from 'svgs/ui/download.svg?sprite';
 
-import Label from 'components/forms/label';
 import { ScenariosSolutionsDetailsProps } from './types';
-import SolutionsTable from '../table';
-import { SolutionRow } from '../table/types';
+import SolutionsTableForm from '../table-form/component';
 
 export const ScenariosSolutionsDetails: React.FC<ScenariosSolutionsDetailsProps> = ({
   onChangeSection,
@@ -24,49 +19,6 @@ export const ScenariosSolutionsDetails: React.FC<ScenariosSolutionsDetailsProps>
   numberOfSchedules,
 }: ScenariosSolutionsDetailsProps) => {
   const [showTable, setShowTable] = useState<boolean>(false);
-
-  const body: SolutionRow[] = [
-    {
-      run: 1,
-      score: 170,
-      cost: 168,
-      'view-on-map': false,
-      best: false,
-      id: 'row1',
-      planningUnits: 168,
-      missingValues: 2,
-    },
-    {
-      run: 2,
-      score: 150,
-      cost: 48,
-      'view-on-map': false,
-      best: true,
-      id: 'row2',
-      planningUnits: 168,
-      missingValues: 2,
-    },
-    {
-      run: 3,
-      score: 110,
-      cost: 18,
-      'view-on-map': false,
-      best: false,
-      id: 'row3',
-      planningUnits: 168,
-      missingValues: 2,
-    },
-    {
-      run: 4,
-      score: 140,
-      cost: 188,
-      'view-on-map': false,
-      best: false,
-      id: 'row4',
-      planningUnits: 168,
-      missingValues: 2,
-    },
-  ];
 
   return (
     <motion.div
@@ -119,54 +71,14 @@ export const ScenariosSolutionsDetails: React.FC<ScenariosSolutionsDetailsProps>
         <Modal
           open={showTable}
           title="Solutions table"
-          size="default"
+          size="wide"
           dismissable
           onDismiss={() => setShowTable(false)}
-          className="text-gray-800"
         >
-          <div className="px-8 pb-8">
-            <div className="flex items-center justify-start pb-6">
-              <Button
-                theme="secondary"
-                size="base"
-                className="flex items-center justify-between pl-4 pr-4"
-                onClick={() => console.info('click - download solutions')}
-              >
-                Download solutions
-                <Icon icon={DOWNLOAD_SVG} className="w-5 h-5 ml-8" />
-              </Button>
-            </div>
-            <div className="flex items-center">
-              <Checkbox
-                theme="light"
-                id="checkbox-5-dif-solutions"
-                className="block w-4 h-4 text-green-300 form-checkbox-dark"
-                onChange={() => console.info('click - 5 most different solutions')}
-              />
-              <Label className="ml-2 text-gray-700">
-                View 5 most different solutions
-              </Label>
-              <button
-                className="ml-3 bg-gray-500 rounded-full opacity-80 p-0.5"
-                type="button"
-              >
-                <Icon icon={INFO_SVG} className="w-4 h-4 text-white" />
-              </button>
-            </div>
-          </div>
-          <SolutionsTable
-            body={body}
-            onSelectSolution={(solution) => console.info('solution selected', solution)}
+          <SolutionsTableForm
+            onCancel={() => setShowTable(false)}
+            onSave={() => setShowTable(false)}
           />
-          <div className="flex items-center justify-center w-full pt-8">
-            <Button
-              theme="primary"
-              size="lg"
-              onClick={() => setShowTable(false)}
-            >
-              Done
-            </Button>
-          </div>
         </Modal>
       </div>
     </motion.div>
