+++ app/layout/scenarios/sidebar/solutions/table-form/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
