+++ app/layout/scenarios/sidebar/solutions/table-form/types.ts
@@ -1,0 +1,4 @@
+export interface SolutionsTableFormProps {
+  onCancel: () => void;
+  onSave: () => void;
+}
